printf: Fix formatting errors
This change fixes the printf formatting errors found by the
compile-time prinf format checker. The errors fall into a few
categories:

1. Incorrect size specifier (missing or extra l).
2. Missing or extra arguments.
3. Bad line splitting.

BUG=chromium:984041
TEST=make -j buildall
BRANCH=none

+++ test/motion_sense_fifo.c
@@ -35,8 +35,8 @@ static int test_insert_async_event(void)
 		sizeof(data), CONFIG_ACCEL_FIFO_SIZE,
 		data, &data_bytes_read);
 	TEST_EQ(read_count, 2, "%d");
-	TEST_EQ(data_bytes_read,
-		2 * sizeof(struct ec_response_motion_sensor_data), "%d");
+	TEST_EQ((size_t)data_bytes_read,
+		2 * sizeof(struct ec_response_motion_sensor_data), "%zd");
 
 	TEST_BITS_SET(data[0].flags, ASYNC_EVENT_FLUSH);
 	TEST_BITS_CLEARED(data[0].flags, MOTIONSENSE_SENSOR_FLAG_ODR);
